downgrade watchify to polling mode
We have problems with the current watchify work.
There are moments when watcher or does not collect the latest revision
of a file in the portion of the stored,
or it just stops working. And this situation is terribly annoying.

This fix enables polling by default.
Polling interval can be set via WATHIFY_POLL_INTERVAL env variable.
You can also overrify completely any settings for watchify
using `.watchifyrc` file in the root of the project.