Move new credit card into it's own component
Also, refactor credit card form to manage display state. Things were
getting to be a little cray when trying to decide what parts
of the form should be shown, and controlling that over the customer
details page, order details page, and new cc vs editing a cc. So,
cleaned it up a bit.