Store SKU code in the form/shadow for a SKU
Moved the actual storage of the SKU code to the form and shadow so that
it's both changeable in the future and versioned. I left the code on the
Sku object for now, so that we can search and index on it in PG. This
decision likely gets refined when we move to generic ObjectHead.