Bug fix for creating a new sku. (#1096)
When you create a new sku, the response had an array as value if several
properties had the same value. This is because the keys in form form would be
duplicated.

The illuminate algorithm would then search for the key but get an array instead of
one value back since multiple keys matched.

The form should not have duplicate keys. This commit fixes that.

Fortunately the data in the database did not have duplicate keys because we
are using jsonb. One limitation of jsonb is that keys are unique and there are no
duplicates, so the data in the DB was actually fine.

But the response to ashes was not.