Certain search views are now scoped.
Each scope gets its own index. When a scope is created the index and mapping
tables should be created via the Scope Processor.

Still need to add nginx endpoint transformations so that ashes transparently
searches the correct index. This will require parsing JWT in lua and appending
the scope to get the correct index.

The challenge is partitioning which search views are scoped and which are not.
In the end, all scope views will be scoped!