First pass on issue with customer routes
The problem is how Akka works with routes: it might be *not* sufficient
to only guard route segment with auth, prefix might be needed. Here I'm
tacking the issue. `/admin` prefix addition is not final, I'm going to
try removing it since I solved some issues I had with tests.