Fixes consumer rebalance problem.
For topics with lots of messages, for example regions, the consumer heartbeat
is not sent often enough because the consumer.poll returns so many messages it
takes a while to process them before the next poll and commmit..

For example, the regions topic kept timing out causing the regions to be indexed
over and over again.

This fix pauses the consumer after the poll and then we send a heartbeat for each
processed message and resume at the end.