Handle hash collisions in an object form.
Use an encoded open hashing attribute key to handle hashing
collisions in an object form. While hash collisions should be rare
within an object form, the change of some object having a collision
reaches 1% after around 10k objects. To handle this we will append an index
to the hash key if the hash is the same but the content is different.

So a hash like "ac3b4" becomes "ac3b3/1" if there is a collision or
"ac3b3/2" and so on.