Add SKU code to the variant <-> sku mapping
Since we're currently going the route where SKU code will be immutable
until we push SKU ID more completely through the system, cache the SKU
code in the variant to SKU mapping table because we have a guarantee
that the code won't change.

Also, do a little bit of renaming so that 'Mwh' isn't appended to every
mention of SKU. Since the SKU only exists in MWH, this is a bit
redundant.