Add a parent ID to object context
This is the genesis of a concept that will allow forking and branching.
The essential concept is that when attempting to find an object in a
context, if it is not found, the parent will be used. This allows us to
do some interesting things - so, branching a context will _not_ require
all objects in the context to be branched.