Update configuration for Azure (#2336)
* Remove failure constraint from go-path.sh

There are errors in the default Azure Ubuntu 17.10 bash profile
configuration. Removing this like prevents the error from borking
the ability to SSH in.

* Mesos configuration should use public docker

- Remove references to docker.foxcommerce.com
- Mesos configuration files should no longer expect port 5000 for
  docker
- Mesos configuration should account for docker permissions

* Build docker images against support Ubuntu

Replaced old 16.10 references in dockerfiles with new 17.10 references

* Build appliance base on Debian Jessie

It's the most recent stable Debian build that supports all our
software. Once we move from Mesos to Kubernetes we'll be in the clear
to move to Debian Stretch.